feat: implement RTL support in UI components and update amount data types in backend

+++ backend/src/main/java/uit/ac/ma/est/kessabpro/controllers/TransactionController.java
@@ -22,7 +22,7 @@ public class TransactionController {
     }
 
     @GetMapping("/amounts/{saleId}")
-    public List<BigDecimal> getTransactionAmounts(@PathVariable UUID saleId) {
+    public List<Double> getTransactionAmounts(@PathVariable UUID saleId) {
         return transactionService.testFindAmountsBySaleId(saleId);
     }
